Resurrect mnist examples. Reenabled tests.
