+++ .travis.yml
@@ -58,20 +58,18 @@ script:
   # Next pytest invoke will run boxed tests, but read the synthetic datasets from cache and not
   # waste time on recreating them
 
+  # Tensorflow and pytorch tests confict (segfault). Split into separate runs.
+  - pytest -v examples/mnist/tests/test_pytorch_mnist.py --cov=./ --cov-append
+
   # Ignore two pytorch tests to prevent static-TLS-caused torch-import seg-fault
-  - pytest -Y --cache-clear -m "not forked" -v --cov=./ --trace-config
-    --ignore=examples/mnist/tests/test_generate_mnist_dataset.py
+  - pytest -Y --cache-clear -m "not forked" -v --cov=./  --cov-append --trace-config
+    --ignore=examples/mnist/tests/test_pytorch_mnist.py
     --ignore=petastorm/tests/test_pytorch_utils.py petastorm examples
 
   # We have a separate run for forked: make sure fixtures are reused as much as possible
   - pytest -Y -m "forked" --forked -v --cov=./ --cov-append --trace-config
-    --ignore=examples/mnist/tests/test_generate_mnist_dataset.py
+    --ignore=examples/mnist/tests/test_pytorch_mnist.py
     --ignore=petastorm/tests/test_pytorch_utils.py petastorm examples
-
-  # Run the pytorch tests separately, in this order, but accumulate code coverage data
-  # Temporary disabled until we figure out (a) segfaults in the first test with `import torch`,
-  # and (b) hanging in `test_read_mnist_dataset`
-  # - pytest -v examples/mnist/tests/test_generate_mnist_dataset.py petastorm/tests/test_pytorch_utils.py --cov=./ --cov-append
 
 after_success:
   - codecov
