+++ examples/mnist/pytorch_example.py
@@ -59,15 +59,15 @@ def train(model, device, train_loader, log_interval, optimizer, epoch):
         loss.backward()
         optimizer.step()
         if batch_idx % log_interval == 0:
-            print('Train Epoch: {} [{}/{} ({:.0f}%)]\tLoss: {:.6f}'.format(
-                epoch, batch_idx * len(data), len(train_loader.reader),
-                100. * batch_idx / len(train_loader), loss.item()))
+            print('Train Epoch: {} [{}]\tLoss: {:.6f}'.format(
+                epoch, batch_idx * len(data), loss.item()))
 
 
 def test(model, device, test_loader):
     model.eval()
     test_loss = 0
     correct = 0
+    count = 0
     with torch.no_grad():
         for data, target in test_loader:
             data, target = data.to(device), target.to(device)
@@ -75,11 +75,11 @@ def test(model, device, test_loader):
             test_loss += F.nll_loss(output, target, reduction='sum').item()  # sum up batch loss
             pred = output.max(1, keepdim=True)[1]  # get the index of the max log-probability
             correct += pred.eq(target.view_as(pred)).sum().item()
+            count += data.shape[0]
 
-    test_loss /= len(test_loader.reader)
+    test_loss /= count
     print('\nTest set: Average loss: {:.4f}, Accuracy: {}/{} ({:.0f}%)\n'.format(
-        test_loss, correct, len(test_loader.reader),
-        100. * correct / len(test_loader.reader)))
+        test_loss, correct, count, 100. * correct / count))
 
 
 def _transform_row(mnist_row):
@@ -116,7 +116,7 @@ def main():
                         help='learning rate (default: 0.01)')
     parser.add_argument('--momentum', type=float, default=0.5, metavar='M',
                         help='SGD momentum (default: 0.5)')
-    parser.add_argument('--no-cuda', action='store_true', default=False,
+    parser.add_argument('--no-cuda', action='store_true', default=True,
                         help='disables CUDA training')
     parser.add_argument('--seed', type=int, default=1, metavar='S',
                         help='random seed (default: 1)')
